Renamed db to database
