+++ middleware/jwt.go
@@ -7,7 +7,7 @@ import (
 
 	jwt "github.com/appleboy/gin-jwt/v2"
 	"github.com/gin-gonic/gin"
-	database "speedrun.website/db"
+	"speedrun.website/database"
 	"speedrun.website/graph/model"
 	"speedrun.website/utils"
 )
